Second exercise
Adds second exercise solutions - finding if a String is a permutaion of
another.

Extracted a test runner as an 'in a pinch'  AssertJ replacement.

Included Java error log, to celebrate the fact of getting a Java core
dump after 20 years of trying. Stars aligned, and trying to generate 12!
strings of length 12 caused the dump instead of throwing
OutOfMemoryError. Wasn't able to reproduce it (getting OOME now).